Use Double prices; add UA header & end-time warning
Multiple changes to support fractional prices, add a User-Agent URLProtocol, and improve end-time validation/UI.

Key changes:
- Replace Int prices with Double across models and viewmodels (TRPTourProduct, mappers, selection/filter VCs, saved plans, activity listing and time selection) to preserve decimal precision.
- Add TRPUserAgentURLProtocol to inject a custom User-Agent for registered hosts and register it during TRPCoreKit startup.
- Add setErrorState to TRPSelectionField for toggling error styling.
- Activity/listing updates: use Double prices, exclude free tours by setting params.minPrice = 1, adjust sorting to handle nil prices using .greatestFiniteMagnitude.
- Filter VC: migrate FilterData price fields to Double and wire slider values without lossy Int conversions.
- Add localization key for end-time-before-start warning.
- Add end-time warning UI in AddPlanTimeAndTravelersVC (warning row, constraints, dynamic content height) and robust handling of midnight picks by snapping stored end time to 23:59 while displaying "00:00".
- Time selection VM: seed selectedDate from segment/step saved date (YMD matching), propagate monetary Double prices, add helpers (extractYMD, findAvailableDay) and auto-apply editing time slot once schedule loads.
- SavedPlansViewModel: return Double price from API item instead of converting to Int.
- Timeline availability: use tripProfile.segments as source of truth, re-run processTimelineData after availability check to refresh snapshot flags.

These changes aim to preserve price precision, improve networking headers, and provide clearer UX/validation for time selection.

+++ TRPCoreKit/TRPDataLayer/Domain/Mappers/Tour/TourMapper.swift
@@ -43,12 +43,13 @@ final class TourMapper {
         // Convert duration from Double (minutes) to Int
         let duration = restModel.duration != nil ? Int(restModel.duration!) : nil
 
-        // Convert price: prefer currentPrice over price, convert to Int
-        let price: Int?
+        // Prefer currentPrice over price; preserve decimal precision so the
+        // UI can show the exact fractional value the API returned.
+        let price: Double?
         if let currentPrice = restModel.currentPrice {
-            price = Int(currentPrice)
+            price = Double(currentPrice)
         } else if let regularPrice = restModel.price {
-            price = Int(regularPrice)
+            price = Double(regularPrice)
         } else {
             price = nil
         }
